presentation ui - mailbox - infrastructure

+++ src/presentation/legallead.desktop.core/implementations/MailPersistence.cs
@@ -9,12 +9,9 @@ namespace legallead.desktop.implementations
         public void Clear()
         {
             var fileName = MailFile;
-            if (string.IsNullOrEmpty(fileName)) { return; }
-            lock (sync)
-            {
-                File.Delete(fileName);
-                File.WriteAllText(fileName, string.Empty);
-            }
+            var folderName = MailSubFolder;
+            ClearFileContent(fileName);
+            ClearChildContent(folderName);
         }
 
         public void Save(string json)
@@ -25,7 +22,22 @@ namespace legallead.desktop.implementations
             {
                 var array = Encoding.UTF8.GetBytes(json);
                 var content = Convert.ToBase64String(array);
-                File.Delete(fileName);
+                if (File.Exists(fileName)) File.Delete(fileName);
+                File.WriteAllText(fileName, content);
+            }
+        }
+        public void Save(string id, string json)
+        {
+            var folderName = MailSubFolder;
+            if (string.IsNullOrEmpty(folderName)) { return; }
+            if (!Guid.TryParse(id, out var indx)) { return; }
+            var suffix = $"{indx:D}.txt";
+            var fileName = Path.Combine(folderName, suffix);
+            lock (sync)
+            {
+                var array = Encoding.UTF8.GetBytes(json);
+                var content = Convert.ToBase64String(array);
+                if (File.Exists(fileName)) File.Delete(fileName);
                 File.WriteAllText(fileName, content);
             }
         }
@@ -50,14 +62,40 @@ namespace legallead.desktop.implementations
             }
         }
 
+        public string? Fetch(string id)
+        {
+            var folderName = MailSubFolder;
+            if (string.IsNullOrEmpty(folderName)) { return null; }
+            if (!Guid.TryParse(id, out var indx)) { return null; }
+            var suffix = $"{indx:D}.txt";
+            var fileName = Path.Combine(folderName, suffix);
+            if (string.IsNullOrEmpty(fileName) || !File.Exists(fileName)) { return null; }
+            lock (sync)
+            {
+                try
+                {
+                    var content = File.ReadAllText(fileName);
+                    var array = Convert.FromBase64String(content);
+                    var converted = Encoding.UTF8.GetString(array);
+                    return converted;
+                }
+                catch (Exception)
+                {
+                    return string.Empty;
+                }
+            }
+        }
+
         private static readonly object sync = new();
         private static string AppFolder => appFolder ??= GetFolder();
         private static string MailFolder => mailFolder ??= GetMailFolder();
+        private static string MailSubFolder => mailSubFolder ??= GetMailSubFolder();
         private static string MailFile => mailFile ??= GetMailFile();
 
         private static string? appFolder;
         private static string? mailFolder;
         private static string? mailFile;
+        private static string? mailSubFolder;
         private static string GetFolder()
         {
             var exeName = Assembly.GetExecutingAssembly().Location;
@@ -75,6 +113,15 @@ namespace legallead.desktop.implementations
             if (!Directory.Exists(child)) return string.Empty;
             return child;
         }
+        private static string GetMailSubFolder()
+        {
+            const string folderName = "_letters";
+            var parent = MailFolder;
+            if (!Directory.Exists(parent)) return string.Empty;
+            var child = Path.Combine(parent, folderName);
+            if (!Directory.Exists(child)) Directory.CreateDirectory(child);
+            return child;
+        }
         private static string GetMailFile()
         {
             const string folderName = "user-data.txt";
@@ -83,6 +130,24 @@ namespace legallead.desktop.implementations
             var child = Path.Combine(parent, folderName);
             if (!File.Exists(child)) return string.Empty;
             return child;
+        }
+        private static void ClearFileContent(string fileName)
+        {
+            if (string.IsNullOrEmpty(fileName)) { return; }
+            lock (sync)
+            {
+                File.Delete(fileName);
+                File.WriteAllText(fileName, string.Empty);
+            }
+        }
+        private static void ClearChildContent(string folderName)
+        {
+            if (string.IsNullOrEmpty(folderName)) { return; }
+            lock (sync)
+            {
+                var di = new DirectoryInfo(folderName);
+                di.Delete(true);
+            }
         }
     }
 }
